AA-820:Upgrade Routing
