+++ src/app/main-dashboard/main-dashboard.module.ts
@@ -1,11 +1,14 @@
-import { NgModule }       from '@angular/core';
-import { CommonModule }   from '@angular/common';
-import { FormsModule }    from '@angular/forms';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { dashboardRouting } from './main-dashboard-routing';
 import { MainDashboardComponent } from './main-dashboard.component';
 import { ClinicDashboardComponent } from '../clinic-dashboard/clinic-dashboard.component';
 import { PatientDashboardComponent } from '../patient-dashboard/patient-dashboard.component';
+
+
+
 @NgModule({
   imports: [
     CommonModule,
