+++ src/app/patient-dashboard/patient-dashboard.module.ts
@@ -1,10 +1,11 @@
-import { NgModule }       from '@angular/core';
-import { CommonModule }   from '@angular/common';
-import { FormsModule }    from '@angular/forms';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { patientDashboardRouting } from './patient-dashboard-routing';
 import { PatientInfoComponent } from './patient-info/patient-info.component';
 import { PatientEncountersComponent } from './patient-encounters/patient-encounters.component';
 import { PatientSearchComponent } from './patient-search/patient-search.component';
+import { PatientDashboardGuard } from './patient-dashboard.guard'
 @NgModule({
   imports: [
     CommonModule,
@@ -17,7 +18,7 @@ import { PatientSearchComponent } from './patient-search/patient-search.componen
     PatientSearchComponent
   ],
   providers: [
-
+    PatientDashboardGuard
   ]
 })
 export class PatientDashboardModule { }
