--- src/app/version.json
@@ -1,7 +1,0 @@
-{
-	"version" : {
-		"name":      "Ampath_Point_of_Care",
-		"buildDate": "Wed Oct 12 2016 17:36:00 GMT+0300 (EAT)",
-		"version":   "5.0.5"
-	}
-}
